+++ src/com/ray/model/entities/Categoria.java
@@ -8,7 +8,7 @@ public class Categoria implements Serializable{
 
     private static final long serialVersionUID = 1L;
     
-    private Integer id;
+    private Long id;
     private String name;
     private User user;
     private Double orcamento;
@@ -18,17 +18,17 @@ public class Categoria implements Serializable{
     public Categoria () {};
     
     
-    public Categoria(Integer id, String name, User user) {
+    public Categoria(Long id, String name, User user) {
 	this.id = id;
 	this.name = name;
 	this.user = user;
     }  
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
